--- .env.local.example
@@ -1,19 +1,0 @@
-NEXT_PUBLIC_SUPABASE_URL=https://xyzcompany.supabase.co
-NEXT_PUBLIC_SUPABASE_ANON_KEY=your_supabase_anon_key
-NEXT_PUBLIC_REDIRECT_URL=http://localhost:3000 # or the production URL (X) replace by NEXT_PUBLIC_VERCEL_URL
-NEXT_PUBLIC_POSTHOG_KEY=key
-NEXT_PUBLIC_POSTHOG_HOST=https://us.i.posthog.com
-VERCEL_ENV=local
-# PearAI server, must run locally
-PEARAI_SERVER_URL=http://localhost:8000
-
-# Analytics
-NEXT_PUBLIC_POSTHOG_KEY="your project api key"
-NEXT_PUBLIC_POSTHOG_HOST=https://us.i.posthog.com
-
-
-# The below variable is automatically set by Vercel for dev and preview environments
-# for dynamic urls. (https://supabase.com/docs/guides/auth/redirect-urls#vercel-preview-urls)
-# process?.env?.NEXT_PUBLIC_VERCEL_URL 
-
-NEXT_PUBLIC_TEST_MODE_ENABLED=true # Set to true if running locally, false otherwise
